.history/backend/requirements_20250612000817.txt
flask
flask-cors
python-dotenv
google-cloud-speech
google-cloud-texttospeech
google-generativeai